review: fixnew17

+++ backend/controllers/cards.js
@@ -15,7 +15,7 @@ const deleteCard = async (req, res, next) => {
   try {
     const { id } = req.params;
     const card = await Card.findOne({ _id: id });
-    if (card) {
+    if (card && id === req.user.id) {
       await Card.deleteOne(card);
       return res.status(200).send({ message: 'Карточка удалена' });
     }
